feat: allow custom font-sizes
to not only to rely on sizes from the modular scale set, a custom
font-size can be passed to the `fluidms-font-size()` mixin to allow
to “break out” of the modular scale in special cases.

BREAKING CHANGE: the name of the `$ms` argument for the
`fluidms-font-size()` mixin has changed to `$font-size` to make
it more natural to use if a custom font-size is used.